+++ lib/instabug_http_client.dart
@@ -14,7 +14,11 @@ import 'package:instabug_http_client/instabug_http_logger.dart';
 import 'package:meta/meta.dart';
 
 class InstabugHttpClient extends InstabugHttpLogger implements http.Client {
-  InstabugHttpClient() : client = http.Client() {
+  // InstabugHttpClient() : client = http.Client() {
+  //   logger = this;
+  // }
+
+  InstabugHttpClient({http.Client? client}) : client = client ?? http.Client() {
     logger = this;
   }